upgrading to scala 3

+++ app/uk/gov/hmrc/economiccrimelevyenrolment/repositories/SessionRepository.scala
@@ -25,6 +25,7 @@ import uk.gov.hmrc.economiccrimelevyenrolment.models.UserAnswers
 import uk.gov.hmrc.mongo.MongoComponent
 import uk.gov.hmrc.mongo.play.json.PlayMongoRepository
 import uk.gov.hmrc.mongo.play.json.formats.MongoJavatimeFormats
+import org.mongodb.scala.SingleObservableFuture
 
 import java.time.{Clock, Instant}
 import java.util.concurrent.TimeUnit
@@ -46,7 +47,7 @@ class SessionRepository @Inject() (
           Indexes.ascending("lastUpdated"),
           IndexOptions()
             .name("lastUpdatedIdx")
-            .expireAfter(appConfig.mongoTtl, TimeUnit.SECONDS)
+            .expireAfter(appConfig.mongoTtl.toLong, TimeUnit.SECONDS)
         ),
         IndexModel(ascending("internalId"), IndexOptions().name("internalIdIdx").unique(true))
       )
